--- src/main/java/com/ssafy/solvedpick/accounts/service/SigninService.java
@@ -1,37 +1,0 @@
-package com.ssafy.solvedpick.accounts.service;
-
-import com.ssafy.solvedpick.accounts.dto.SignInFormDTO;
-import com.ssafy.solvedpick.jwt.JwtUtil;
-import com.ssafy.solvedpick.members.domain.Member;
-import com.ssafy.solvedpick.members.repository.MemberRepository;
-import com.ssafy.solvedpick.accounts.dto.TokenResponse;
-import com.ssafy.solvedpick.global.error.exception.UserInfoErrorException;
-
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class SigninService {
-    private final MemberRepository memberRepository;
-    private final BCryptPasswordEncoder passwordEncoder;
-    private final JwtUtil jwtUtil;
-
-    public TokenResponse signIn(SignInFormDTO signInFormDTO) {
-    	Member member = memberRepository.findByUsername(signInFormDTO.getUsername())
-                .orElseThrow(() -> new UserInfoErrorException("Login Error"));
-        
-        if (!passwordEncoder.matches(signInFormDTO.getPassword(), member.getPassword())) {
-            throw new UserInfoErrorException("Login Error");
-        }
-
-        String accessToken = jwtUtil.generateAccessToken(signInFormDTO.getUsername());
-
-        TokenResponse tokenResponse = TokenResponse.builder()
-        				.accessToken(accessToken)
-        				.build();
-        return tokenResponse;
-    }
-}
