+++ src/main/java/com/ssafy/solvedpick/auth/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.ssafy.solvedpick.accounts.dto;
+package com.ssafy.solvedpick.auth.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -10,8 +10,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class SignInFormDTO {
-    private String username;
-    private String password;
+public class TokenResponse {
 
+    private String accessToken;
 }
