+++ src/main/java/com/ssafy/solvedpick/auth/repository/VerificationKeyRepository.java
@@ -1,11 +1,11 @@
-package com.ssafy.solvedpick.accounts.repository;
+package com.ssafy.solvedpick.auth.repository;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ssafy.solvedpick.accounts.domain.VerificationKey;
+import com.ssafy.solvedpick.auth.domain.VerificationKey;
 
 public interface VerificationKeyRepository extends JpaRepository<VerificationKey, Long>{
